update comments, put function into reco/utils
